Refactor after merging with develop
