+++ web/src/main/java/com/infoshareacademy/gitloopersi/freemarker/ConfigProvider.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy.gitLoopersi.freemarker;
+package com.infoshareacademy.gitloopersi.freemarker;
 
 import freemarker.template.Configuration;
 import freemarker.template.TemplateExceptionHandler;
