+++ web/src/main/java/com/infoshareacademy/gitloopersi/servlet/admin/VacationManagerServlet.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy.gitLoopersi.servlet.user;
+package com.infoshareacademy.gitloopersi.servlet.admin;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -9,8 +9,8 @@ import javax.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-@WebServlet("/vacation")
-public class MyVacationServlet extends HttpServlet {
+@WebServlet("/admin/vacation")
+public class VacationManagerServlet extends HttpServlet {
 
   private Logger logger = LoggerFactory.getLogger(getClass().getName());
